+++ src/autd3/benches/gain/primitive.rs
@@ -4,7 +4,7 @@
  * Created Date: 31/07/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 31/07/2023
+ * Last Modified: 18/08/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -13,6 +13,7 @@
 
 mod helper;
 
+use autd3_core::gain::GainFilter;
 use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
 
 use autd3::{core::gain::Gain, prelude::*};
@@ -33,7 +34,7 @@ fn focus(c: &mut Criterion) {
                         black_box(70.),
                         black_box(150.),
                     ))
-                    .calc(geometry)
+                    .calc(geometry, GainFilter::All)
                     .unwrap();
                 })
             },
@@ -48,7 +49,7 @@ fn focus(c: &mut Criterion) {
                         black_box(70.),
                         black_box(150.),
                     ))
-                    .calc(geometry)
+                    .calc(geometry, GainFilter::All)
                     .unwrap();
                 })
             },
@@ -67,14 +68,13 @@ fn focus_cached(c: &mut Criterion) {
             black_box(70.),
             black_box(150.),
         ))
-        .with_cache(&geometry)
-        .unwrap();
+        .with_cache();
         group.bench_with_input(
             BenchmarkId::new("Legacy", size * size),
             &geometry,
             |b, geometry| {
                 b.iter(|| {
-                    g.calc(geometry).unwrap();
+                    g.calc(geometry, GainFilter::All).unwrap();
                 })
             },
         );
@@ -84,14 +84,13 @@ fn focus_cached(c: &mut Criterion) {
             black_box(70.),
             black_box(150.),
         ))
-        .with_cache(&geometry)
-        .unwrap();
+        .with_cache();
         group.bench_with_input(
             BenchmarkId::new("Advanced", size * size),
             &geometry,
             |b, geometry| {
                 b.iter(|| {
-                    g.calc(geometry).unwrap();
+                    g.calc(geometry, GainFilter::All).unwrap();
                 })
             },
         );
@@ -113,7 +112,7 @@ fn bessel(c: &mut Criterion) {
                         Vector3::new(black_box(0.), black_box(0.), black_box(1.)),
                         black_box(0.1),
                     )
-                    .calc(geometry)
+                    .calc(geometry, GainFilter::All)
                     .unwrap();
                 })
             },
@@ -128,7 +127,7 @@ fn bessel(c: &mut Criterion) {
                         Vector3::new(black_box(0.), black_box(0.), black_box(1.)),
                         black_box(0.1),
                     )
-                    .calc(geometry)
+                    .calc(geometry, GainFilter::All)
                     .unwrap();
                 })
             },
@@ -147,7 +146,7 @@ fn plane(c: &mut Criterion) {
             |b, geometry| {
                 b.iter(|| {
                     Plane::new(Vector3::new(black_box(0.), black_box(0.), black_box(1.)))
-                        .calc(geometry)
+                        .calc(geometry, GainFilter::All)
                         .unwrap();
                 })
             },
@@ -158,7 +157,7 @@ fn plane(c: &mut Criterion) {
             |b, geometry| {
                 b.iter(|| {
                     Plane::new(Vector3::new(black_box(0.), black_box(0.), black_box(1.)))
-                        .calc(geometry)
+                        .calc(geometry, GainFilter::All)
                         .unwrap();
                 })
             },
@@ -167,5 +166,72 @@ fn plane(c: &mut Criterion) {
     group.finish();
 }
 
-criterion_group!(benches, focus, focus_cached, bessel, plane);
+fn group_dev(c: &mut Criterion) {
+    let mut group = c.benchmark_group("group");
+
+    group.bench_with_input(
+        "group-dev",
+        &generate_geometry::<LegacyTransducer>(3),
+        |b, geometry| {
+            b.iter(|| {
+                (0..geometry.num_devices())
+                    .fold(Group::by_device(|dev| Some(dev)), |acc, i| {
+                        acc.set(
+                            i,
+                            Focus::new(Vector3::new(
+                                black_box(90.),
+                                black_box(70.),
+                                black_box(150.),
+                            )),
+                        )
+                    })
+                    .calc(geometry, GainFilter::All)
+                    .unwrap();
+            })
+        },
+    );
+    group.finish();
+}
+
+fn group_trans(c: &mut Criterion) {
+    let mut group = c.benchmark_group("group");
+
+    group.bench_with_input(
+        "group-trans",
+        &generate_geometry::<LegacyTransducer>(3),
+        |b, geometry| {
+            b.iter(|| {
+                (0..geometry.num_devices())
+                    .fold(
+                        Group::<_, LegacyTransducer>::by_transducer(|tr| {
+                            Some(tr.idx() / autd3_core::autd3_device::AUTD3::NUM_TRANS_IN_UNIT)
+                        }),
+                        |acc, i| {
+                            acc.set(
+                                i,
+                                Focus::new(Vector3::new(
+                                    black_box(90.),
+                                    black_box(70.),
+                                    black_box(150.),
+                                )),
+                            )
+                        },
+                    )
+                    .calc(geometry, GainFilter::All)
+                    .unwrap();
+            })
+        },
+    );
+    group.finish();
+}
+
+criterion_group!(
+    benches,
+    focus,
+    focus_cached,
+    bessel,
+    plane,
+    group_dev,
+    group_trans
+);
 criterion_main!(benches);