+++ src/autd3/src/gain/mod.rs
@@ -14,7 +14,7 @@
 pub mod bessel;
 pub mod cache;
 pub mod focus;
-pub mod grouped;
+pub mod group;
 pub mod null;
 pub mod plane;
 pub mod trans_test;
@@ -23,7 +23,7 @@ pub mod uniform;
 pub use bessel::Bessel;
 pub use cache::Cache;
 pub use focus::Focus;
-pub use grouped::Grouped;
+pub use group::Group;
 pub use null::Null;
 pub use plane::Plane;
 pub use trans_test::TransducerTest;