+++ src/examples/src/tests/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 28/05/2021
  * Author: Shun Suzuki
  * -----
- * Last Modified: 30/05/2023
+ * Last Modified: 19/08/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2021 Shun Suzuki. All rights reserved.
@@ -16,7 +16,7 @@ mod bessel;
 mod custom;
 mod flag;
 mod focus;
-mod grouped;
+mod group;
 mod holo;
 mod plane;
 mod stm;
@@ -27,7 +27,7 @@ pub use bessel::*;
 pub use custom::*;
 pub use flag::*;
 pub use focus::*;
-pub use grouped::*;
+pub use group::*;
 pub use holo::*;
 pub use plane::*;
 pub use stm::*;